==> src/03_alphadiversity.Rmd <==
---
title: "Alpha,Beta analysis"
author: "Adrià Auladell Martín"
output:
  pdf_document: 
    highlight: zenburn
  html_document: default
---

#Packages
```{r, message=FALSE, warning=FALSE}
library(vegan)
library(tidyverse)
library(phyloseq)
library(ggrepel)
library(mgcv)
```

# Output of the script

```{r}

dir.create(file.path('../results/figures/alphadiv'), showWarnings = FALSE)

```

#Data import

```{r}
bl.phy <- readRDS("../data/robjects/blphyloseq.rds") 
source("sourcefiles/params-graphs.R") 
source("sourcefiles/main_functions.R")

physeq.list <- data_transformation(bl.phy)

bl.phy.relab <- physeq.list$relab
bl.phy.raref <- physeq.list$raref
bl.phy.log <- physeq.list$log
bl.phy.clr <- physeq.list$clr

# since the list is quite big let's erase it
physeq.list <- NULL
```

## Fractions
```{r}
bl.phy.0.2 <- subset_samples(bl.phy, fraction == '0.2') %>% 
  prune_taxa(taxa = taxa_sums(.) != 0, .)

bl.phy.3 <- subset_samples(bl.phy, fraction == '3') %>% 
  prune_taxa(taxa = taxa_sums(.) != 0, .)
```

# Alpha diversity 

```{r}
rich.df <- data.frame(estimate_richness(bl.phy.raref),
                      sample_data(bl.phy))

write_tsv(rich.df, 
          '../results/summary_statistics/01_richness-df-rarefied.tsv')

richall.df <- data.frame(estimate_richness(bl.phy),
                         sample_data(bl.phy))
write_tsv(richall.df, 
          '../results/summary_statistics/01_richness-df-withoutraref.tsv')

```


# Time + fraction 

## Data vizualization
```{r}
levels <- interaction(rich.df$season, rich.df$fraction) %>% unique

cols <- c("#63bfb8", "darkslategray4" , "#52b216",
          "darkgreen" , "khaki1", "yellow4" , "#e3a40c", "darkorange3") 

names(cols) <- levels


rich.plot <- c('Chao1', 'Shannon', 'Simpson') %>% 
  map(~ ggplot(rich.df, aes_string('month', .x)) +
        geom_point(aes(group = interaction(season,fraction)),
                   position = position_dodge(width = 1),
                   color = 'grey', alpha = 0.5) +
        geom_boxplot(aes(fill=interaction(season,fraction)),
                     outlier.colour = NA,
                     fatten = 1) +
        scale_x_month + 
        theme_bw() + 
        scale_fill_manual(values = cols,
                          limits = as.character(levels)) +
        guides(fill = F)
  )

filenames <- str_c( "../results/figures/alphadiv/byfraction_",
                    c('Chao1', 'Shannon', 'Simpson'),
                    "_raref.pdf")


map2(filenames, rich.plot,  ggsave, height = 4, width = 7)

library(patchwork)


composite <- rich.plot[[1]] +
  theme(axis.title.x = element_blank()) +
  rich.plot[[2]] + 
  plot_layout(ncol = 1) + 
  plot_annotation(tag_levels = 'A')



ggsave(filename = '../results/figures/alphadiv/byfraction_chao_shannon.pdf',
       width = 8, height = 6)
```


### Statistical differences?

```{r}

rich.kruskal <- rich.df %>% 
  mutate(season.two = case_when(season == "spring"~"summer",
                                season == "autumn"~ "winter",
                                TRUE ~ season),
         season = as.factor(season),
         fraction = as.factor(fraction),
         interaction = interaction(season, fraction),
         interaction2 = interaction(season.two, fraction),
         )

pairwise.wilcox.test(rich.kruskal$Shannon, 
                     rich.kruskal$season,
                     p.adjust.method = "fdr")


pairwise.wilcox.test(rich.kruskal$Chao1, 
                     rich.kruskal$season,
                     p.adjust.method = "fdr")


pairwise.wilcox.test(rich.kruskal$Shannon, 
                     rich.kruskal$fraction,
                     p.adjust.method = "fdr")


pairwise.wilcox.test(rich.kruskal$Chao1, 
                     rich.kruskal$fraction,
                     p.adjust.method = "fdr")

pair.all <- pairwise.wilcox.test(rich.kruskal$Chao1, 
                     rich.kruskal$interaction,
                     p.adjust.method = "fdr")

pairwise.wilcox.test(rich.kruskal$Shannon, 
                     rich.kruskal$interaction,
                     p.adjust.method = "fdr")


pairwise.wilcox.test(rich.kruskal$Shannon, 
                     rich.kruskal$interaction2,
                     p.adjust.method = "fdr")

pairwise.wilcox.test(rich.kruskal$Chao1, 
                     rich.kruskal$interaction2,
                     p.adjust.method = "fdr")

pairs.results.tsv <- pair.all$p.value %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "group") %>% 
  gather(key = "group2", value = "pval" , -group) 


pairs.results.tsv %>% 
  filter(pval < 0.05)

write_tsv(pairs.results.tsv,
          '../results/summary_statistics/01_wilcox-allgroups-alphadiv.tsv')

```
